KS-1-4-variables

+++ app/src/main/kotlin/variables/variables4.kt
@@ -3,7 +3,7 @@ package org.example.app.variables
 fun main() {
 
     val lenght: Long = 40_868_600_000
-    val age: Short = 27
+    val age: Byte = 27
     val partOfDay: Float = 0.075f
     val seconds: Short = 6480
     val partOfYear: Double = 2.0547945205479453E-4
